use ! in shell to enter spaces. Add a session command prep method to drain buffers before issuing a command.

+++ smb/source/client/smb2/smb2diagnostics.hpp
@@ -56,7 +56,7 @@ public:
 
   void display_text_warnings()
   {
-     for (int i = 0; i <  warnings.size(); i++)
+     for (int i = 0; i <  (int)warnings.size(); i++)
      {
         diag_printf_fn(DIAG_INFORMATIONAL,"Warning: %s\n", warnings[i]);
         free(warnings[i]);
